Fixed grammar to understand operator precedence. Refactored code.

+++ src/assignment1/parser/operators/relational/GreaterThanOperator.java
@@ -1,6 +1,7 @@
 package assignment1.parser.operators.relational;
 
 import assignment1.parser.operands.BooleanOp;
+import assignment1.parser.operands.IntegerOp;
 import assignment1.parser.operands.Operand;
 
 import java.util.List;
@@ -17,7 +18,7 @@ public class GreaterThanOperator extends RelationalOperator {
     @Override
     public Operand compute() {
 
-        boolean result = (Float) operands.get(0).getValue() > (Float) operands.get(1).getValue();
+        boolean result = (Integer) operands.get(0).getValue() > (Integer) operands.get(1).getValue();
         return BooleanOp.of(result);
     }
